revert: "refactor(cli, lib): run the emulation in a separate thread"
Not worth it to run in a separate thread, what with the added
complexity. Input wasn't responsive, and having to deal with the Tick
event while handling input didn't feel very elegant. Input has been
refactored to use a bool to check if the chip8 is waiting for input, and
if it is, then the next execution cycle doesn't run.

This reverts commit 4a3e62a5bb9eefdbddf637e7a628b1626c8c2f36.